GCC default optimization level is 0, i.e. no optimization. Production environment usually has optimization level at least 2. 
When `-O3` is used, `cqueue` is three times slower than `std::deque`. The root cause is modulo operator. Every push/pop needs to modulo `reserved_`. 
Since `reserved_` is always power of 2. Modulo can be replaced by bit operation.
Before the PR
```
g++ -std=c++20 -g -03 -o deque-prof deque-prof.cpp
g++ -std=c++20 -g -03 -o cqueue-prof cqueue-prof.cpp
./deque-prof 
Elapsed time in microseconds : 63234 us
./cqueue-prof
Elapsed time in microseconds : 224606 us
```
After:
```
./deque-prof 
Elapsed time in microseconds : 63064 us
./cqueue-prof
Elapsed time in microseconds : 33249 us
```