@torrentg 2024-06-22
Impressive improvement !
I slightly modified the suggested change to consider the case mReserved != power of 2.
Thank you for your contribution.