@372046933 2024-06-23
Thank you for your review. Does `shrink_to_fit` break  the power of 2 guarantee? Since growth is always power of 2