@torrentg 2024-06-24
You are welcome. Yes, shrink_to_fit and the constructor with fixed capacity breaks the power of two garantee. This was reported by some tests that failed. I suspect that usual usage case don't use this features, this is because I set a [likely] in the power of two case.